去掉LocalActivityManager，解决与高德地图冲突的问题
