+++ build.gradle
@@ -5,8 +5,8 @@ dependencies {
 }
 
 android {
-    compileSdkVersion 21
-    buildToolsVersion "21.1.2"
+    compileSdkVersion 22
+    buildToolsVersion "23.0.0"
     lintOptions {
         abortOnError false
     }
