+++ uexChatKeyboard/AndroidManifest.xml
@@ -1,14 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest>
 
-    <activity
-        android:name="org.zywx.wbpalmstar.plugin.chatkeyboard.ACEChatKeyboardActivity"
-        android:configChanges="orientation"
-        android:screenOrientation="portrait"
-        android:theme="@android:style/Theme.NoTitleBar.Fullscreen"
-        android:windowSoftInputMode="stateHidden|adjustResize" >
-    </activity>
-
     <uses-permission android:name="android.permission.CAMERA" />
     <uses-permission android:name="android.permission.RECORD_AUDIO" />
     <uses-permission android:name="android.permission.MOUNT_UNMOUNT_FILESYSTEMS" />
